requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Faker==24.3.0
idna==3.6
mysqlclient==2.2.4
pillow==10.2.0
PyJWT==2.8.0
python-benedict==0.33.2
python-dateutil==2.9.0.post0
python-fsutil==0.14.1
python-slugify==8.0.4
PyYAML==6.0.1
requests==2.31.0
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing==3.7.4.3
tzdata==2024.1
urllib3==2.2.1
